requirements.txt
fastapi
uvicorn
google-cloud-bigquery
gunicorn